+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.8 - (In Development)
 
+1. Register the XSLT 3.0 schema file URI with IntelliJ.
+
 ## 2020
 
 *  [1.7 - 2020-03-21](docs/_posts/2020-07-10-release-1.7.md)